Moving entity error dialog to util, planning to use it everywhere
