+++ octane-eclipse-plugin/src/com/hpe/octane/ideplugins/eclipse/ui/util/EntityErrorDialog.java
@@ -10,7 +10,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package com.hpe.octane.ideplugins.eclipse.ui.entitydetail;
+package com.hpe.octane.ideplugins.eclipse.ui.util;
 
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.SelectionAdapter;
